refactor: move ui from `no-meteor` to `node-app`
