+++ imports/node-app/plugins/ui/i18n/fr.json
