+++ imports/node-app/plugins/ui/i18n/nb.json
