+++ imports/plugins/core/ui/register.js
@@ -6,6 +6,6 @@
  * temporary bridge.
  */
 import Reaction from "/imports/plugins/core/core/server/Reaction";
-import register from "./server/no-meteor/register";
+import register from "/imports/node-app/plugins/ui/index.js";
 
 Reaction.whenAppInstanceReady(register);
